+++ example/spec/parallel/discoverGranules/DiscoverGranulesHttpsSpec.js
@@ -1,4 +1,3 @@
-const { Execution } = require('@cumulus/api/models');
 const { deleteExecution } = require('@cumulus/api-client/executions');
 const { deleteGranule } = require('@cumulus/api-client/granules');
 const { deleteProvider } = require('@cumulus/api-client/providers');
@@ -9,10 +8,11 @@ const {
   cleanupCollections,
   waitForCompletedExecution,
 } = require('@cumulus/integration-tests');
+const { getExecution } = require('@cumulus/api-client/executions');
 
 const { loadConfig, createTimestampedTestId, createTestSuffix } = require('../../helpers/testUtils');
 const { buildHttpOrHttpsProvider, createProvider } = require('../../helpers/Providers');
-const { waitForModelStatus } = require('../../helpers/apiUtils');
+const { waitForApiStatus } = require('../../helpers/apiUtils');
 
 const workflowName = 'DiscoverGranules';
 
@@ -20,12 +20,11 @@ const workflowName = 'DiscoverGranules';
 
 describe('The Discover Granules workflow with https Protocol', () => {
   const collectionsDir = './data/collections/https_testcollection_001/';
-  let httpsWorkflowExecution = null;
+  let httpsWorkflowExecution;
 
   let collection;
   let config;
   let discoverGranulesLambdaOutput;
-  let executionModel;
   let lambdaStep;
   let provider;
   let queueGranulesOutput;
@@ -36,9 +35,6 @@ describe('The Discover Granules workflow with https Protocol', () => {
 
   beforeAll(async () => {
     config = await loadConfig();
-
-    process.env.ExecutionsTable = `${config.stackName}-ExecutionsTable`;
-    executionModel = new Execution();
 
     testId = createTimestampedTestId(config.stackName, 'DiscoverGranulesHttps');
     testSuffix = createTestSuffix(testId);
@@ -131,9 +127,12 @@ describe('The Discover Granules workflow with https Protocol', () => {
 
   describe('the reporting lambda has received the cloudwatch stepfunction event and', () => {
     it('the execution record is added to DynamoDB', async () => {
-      const record = await waitForModelStatus(
-        executionModel,
-        { arn: httpsWorkflowExecution.executionArn },
+      const record = await waitForApiStatus(
+        getExecution,
+        {
+          prefix: config.stackName,
+          arn: httpsWorkflowExecution.executionArn,
+        },
         'completed'
       );
       expect(record.status).toEqual('completed');