+++ example/spec/parallel/ingestGranule/IngestGranuleSuccessSpec.js
@@ -13,7 +13,6 @@ const intersection = require('lodash/intersection');
 const isObject = require('lodash/isObject');
 
 const {
-  Execution,
   Granule,
   Pdr,
 } = require('@cumulus/api/models');
@@ -65,7 +64,9 @@ const {
   getTEARequestHeaders,
 } = require('@cumulus/integration-tests/api/distribution');
 const { LambdaStep } = require('@cumulus/integration-tests/sfnStep');
+const { getExecution } = require('@cumulus/api-client/executions');
 
+const { waitForApiStatus } = require('../../helpers/apiUtils');
 const {
   loadConfig,
   templateFile,
@@ -116,7 +117,6 @@ describe('The S3 Ingest Granules workflow', () => {
   let beforeAllError;
   let collection;
   let config;
-  let executionModel;
   let expectedPayload;
   let expectedS3TagSet;
   let expectedSyncGranulePayload;
@@ -147,8 +147,6 @@ describe('The S3 Ingest Granules workflow', () => {
 
       process.env.GranulesTable = `${config.stackName}-GranulesTable`;
       granuleModel = new Granule();
-      process.env.ExecutionsTable = `${config.stackName}-ExecutionsTable`;
-      executionModel = new Execution();
       process.env.system_bucket = config.bucket;
       process.env.ProvidersTable = `${config.stackName}-ProvidersTable`;
       process.env.PdrsTable = `${config.stackName}-PdrsTable`;
@@ -302,11 +300,13 @@ describe('The S3 Ingest Granules workflow', () => {
 
   it('triggers a running execution record being added to DynamoDB', async () => {
     if (beforeAllError) throw SetupError;
-
-    const record = await waitForModelStatus(
-      executionModel,
-      { arn: workflowExecutionArn },
-      ['running', 'completed']
+    const record = await waitForApiStatus(
+      getExecution,
+      {
+        prefix: config.stackName,
+        arn: workflowExecutionArn,
+      },
+      ['running']
     );
     expect(record.status).toEqual('running');
   });
@@ -729,9 +729,12 @@ describe('The S3 Ingest Granules workflow', () => {
 
     it('triggers the successful execution record being added to DynamoDB', async () => {
       if (beforeAllError || subTestSetupError) throw SetupError;
-      const record = await waitForModelStatus(
-        executionModel,
-        { arn: workflowExecutionArn },
+      const record = await waitForApiStatus(
+        getExecution,
+        {
+          prefix: config.stackName,
+          arn: workflowExecutionArn,
+        },
         'completed'
       );
       expect(record.status).toEqual('completed');
@@ -739,9 +742,12 @@ describe('The S3 Ingest Granules workflow', () => {
 
     it('triggers the failed execution record being added to DynamoDB', async () => {
       if (beforeAllError || subTestSetupError) throw SetupError;
-      const record = await waitForModelStatus(
-        executionModel,
-        { arn: failingWorkflowExecution.executionArn },
+      const record = await waitForApiStatus(
+        getExecution,
+        {
+          prefix: config.stackName,
+          arn: failingWorkflowExecution.executionArn,
+        },
         'failed'
       );
       expect(record.status).toEqual('failed');
@@ -829,7 +835,6 @@ describe('The S3 Ingest Granules workflow', () => {
       let cmrLink;
       let publishGranuleExecution;
       let updateCmrAccessConstraintsExecutionArn;
-
       let subTestSetupError;
 
       beforeAll(async () => {
@@ -1004,10 +1009,12 @@ describe('The S3 Ingest Granules workflow', () => {
 
         it('saves asyncOperationId to execution record', async () => {
           if (beforeAllError || subTestSetupError) throw SetupError;
-
-          const reingestExecution = await waitForModelStatus(
-            executionModel,
-            { arn: reingestExecutionArn },
+          const reingestExecution = await waitForApiStatus(
+            getExecution,
+            {
+              prefix: config.stackName,
+              arn: reingestExecutionArn,
+            },
             'completed'
           );
           expect(reingestExecution.asyncOperationId).toEqual(asyncOperationId);