+++ example/spec/parallel/testAPI/bulkOperationsSuccessSpec.js
@@ -195,6 +195,10 @@ describe('POST /granules/bulk', () => {
         await deleteAsyncOperation({ prefix: config.stackName, asyncOperationId: postBulkOperationsBody.id });
       }
 
+      if (postBulkOperationsBody.id) {
+        await deleteAsyncOperation({ prefix: config.stackName, asyncOperationId: postBulkOperationsBody.id });
+      }
+
       await pAll(
         [
           () => deleteProvider({ prefix, providerId: get(provider, 'id') }),