+++ example/spec/serial/DiscoverGranulesHttpSpec.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const { Execution } = require('@cumulus/api/models');
 const { LambdaStep } = require('@cumulus/integration-tests/sfnStep');
 const { deleteExecution } = require('@cumulus/api-client/executions');
 const { deleteGranule } = require('@cumulus/api-client/granules');
@@ -12,14 +11,15 @@ const {
   cleanupCollections,
   waitForCompletedExecution,
 } = require('@cumulus/integration-tests');
+const { getExecution } = require('@cumulus/api-client/executions');
 
+const { waitForApiStatus } = require('../helpers/apiUtils');
 const {
   loadConfig,
   createTimestampedTestId,
   createTestSuffix,
 } = require('../helpers/testUtils');
 const { buildHttpOrHttpsProvider, createProvider } = require('../helpers/Providers');
-const { waitForModelStatus } = require('../helpers/apiUtils');
 
 const workflowName = 'DiscoverGranules';
 
@@ -32,7 +32,6 @@ describe('The Discover Granules workflow with http Protocol', () => {
   let discoverGranulesExecution;
   let discoverGranulesExecutionArn;
   let discoverGranulesLambdaOutput;
-  let executionModel;
   let ignoringFilesConfigExecutionArn;
   let ignoringFilesIngestExecutionArns;
   let ingestGranuleWorkflowArn1;
@@ -51,9 +50,6 @@ describe('The Discover Granules workflow with http Protocol', () => {
   beforeAll(async () => {
     try {
       config = await loadConfig();
-
-      process.env.ExecutionsTable = `${config.stackName}-ExecutionsTable`;
-      executionModel = new Execution();
 
       testId = createTimestampedTestId(config.stackName, 'DiscoverGranules');
       testSuffix = createTestSuffix(testId);
@@ -159,9 +155,12 @@ describe('The Discover Granules workflow with http Protocol', () => {
 
   describe('the reporting lambda has received the CloudWatch step function event and', () => {
     it('the execution record is added to DynamoDB', async () => {
-      const record = await waitForModelStatus(
-        executionModel,
-        { arn: discoverGranulesExecutionArn },
+      const record = await waitForApiStatus(
+        getExecution,
+        {
+          prefix: config.stackName,
+          arn: discoverGranulesExecution.executionArn,
+        },
         'completed'
       );
       expect(record.status).toEqual('completed');