+++ packages/api/tests/endpoints/test-rules.js
@@ -15,12 +15,14 @@ const {
   ProviderPgModel,
   translateApiCollectionToPostgresCollection,
   translateApiProviderToPostgresProvider,
-  translateApiRuleToPostgresRule,
 } = require('@cumulus/db');
 const S3 = require('@cumulus/aws-client/S3');
-const { bootstrapElasticSearch } = require('@cumulus/es-client/bootstrap');
 const { Search } = require('@cumulus/es-client/search');
 const indexer = require('@cumulus/es-client/indexer');
+const {
+  createTestIndex,
+  cleanupTestIndex,
+} = require('@cumulus/es-client/testUtils');
 
 const { buildFakeExpressResponse } = require('./utils');
 const {
@@ -29,8 +31,9 @@ const {
   fakeRuleFactoryV2,
   createFakeJwtAuthToken,
   setAuthorizedOAuthUsers,
+  createRuleTestRecords,
 } = require('../../lib/testUtils');
-const { post } = require('../../endpoints/rules');
+const { post, put, del } = require('../../endpoints/rules');
 const AccessToken = require('../../models/access-tokens');
 const Rule = require('../../models/rules');
 const assertions = require('../../lib/assertions');
@@ -53,7 +56,6 @@ const testDbName = randomString(12);
 // import the express app after setting the env variables
 const { app } = require('../../app');
 
-const esIndex = randomString();
 const workflow = randomId('workflow-');
 const testRule = fakeRuleFactoryV2({
   name: randomId('testRule'),
@@ -71,7 +73,6 @@ const dynamoRuleOmitList = ['createdAt', 'updatedAt', 'state', 'provider', 'coll
 
 const setBuildPayloadStub = () => sinon.stub(Rule, 'buildPayload').resolves({});
 
-let esClient;
 let jwtAuthToken;
 let accessTokenModel;
 let ruleModel;
@@ -88,11 +89,15 @@ test.before(async (t) => {
   t.context.testKnex = knex;
   t.context.testKnexAdmin = knexAdmin;
 
-  const esAlias = randomString();
-  process.env.ES_INDEX = esAlias;
-  await bootstrapElasticSearch('fakehost', esIndex, esAlias);
+  const { esIndex, esClient } = await createTestIndex();
+  t.context.esIndex = esIndex;
+  t.context.esClient = esClient;
+  t.context.esRulesClient = new Search(
+    {},
+    'rule',
+    t.context.esIndex
+  );
 
-  esClient = await Search.es('fakehost');
   await S3.createBucket(process.env.system_bucket);
 
   buildPayloadStub = setBuildPayloadStub();
@@ -103,9 +108,10 @@ test.before(async (t) => {
 
   ruleModel = new Rule();
   await ruleModel.createTable();
+  t.context.ruleModel = ruleModel;
 
   const ruleRecord = await ruleModel.create(testRule);
-  await indexer.indexRule(esClient, ruleRecord, esAlias);
+  await indexer.indexRule(esClient, ruleRecord, t.context.esIndex);
 
   const username = randomString();
   await setAuthorizedOAuthUsers([username]);
@@ -114,7 +120,6 @@ test.before(async (t) => {
   await accessTokenModel.createTable();
 
   jwtAuthToken = await createFakeJwtAuthToken({ accessTokenModel, username });
-  esClient = await Search.es('fakehost');
 });
 
 test.beforeEach((t) => {
@@ -128,7 +133,7 @@ test.after.always(async (t) => {
   await accessTokenModel.deleteTable();
   await ruleModel.deleteTable();
   await S3.recursivelyDeleteS3Bucket(process.env.system_bucket);
-  await esClient.indices.delete({ index: esIndex });
+  await cleanupTestIndex(t.context);
 
   buildPayloadStub.restore();
   await destroyLocalTestDb({
@@ -319,7 +324,7 @@ test('403 error when calling the API endpoint to delete an existing rule without
   t.deepEqual(response.body, record);
 });
 
-test('POST creates a rule', async (t) => {
+test('POST creates a rule in all data stores', async (t) => {
   const { newRule } = t.context;
 
   const fakeCollection = fakeCollectionFactory();
@@ -385,6 +390,11 @@ test('POST creates a rule', async (t) => {
       dynamoRuleOmitList
     )
   );
+
+  const esRecord = await t.context.esRulesClient.get(
+    newRule.name
+  );
+  t.like(esRecord, fetchedDynamoRecord);
 });
 
 test('POST creates a rule in Dynamo and PG with correct timestamps', async (t) => {
@@ -549,25 +559,20 @@ test.serial('POST returns a 500 response if record creation throws unexpected er
   }
 });
 
-test.serial('POST does not write to RDS or DynamoDB if writing to RDS fails', async (t) => {
+test.serial('post() does not write to Elasticsearch/DynamoDB if writing to PostgreSQL fails', async (t) => {
   const { newRule, testKnex } = t.context;
 
-  const failingTrx = (cb) => {
-    const fakeTrx = sinon.stub().returns({
-      insert: () => {
-        throw new Error('Insert Rule Error');
-      },
-    });
-    return cb(fakeTrx);
+  const fakeRulePgModel = {
+    create: () => {
+      throw new Error('something bad');
+    },
   };
-
-  const trxStub = sinon.stub(testKnex, 'transaction').callsFake(failingTrx);
-  t.teardown(() => trxStub.restore());
 
   const expressRequest = {
     body: newRule,
     testContext: {
-      dbClient: testKnex,
+      knex: testKnex,
+      rulePgModel: fakeRulePgModel,
     },
   };
   const response = buildFakeExpressResponse();
@@ -576,26 +581,30 @@ test.serial('POST does not write to RDS or DynamoDB if writing to RDS fails', as
   const dbRecords = await t.context.rulePgModel
     .search(t.context.testKnex, { name: newRule.name });
 
-  t.true(response.boom.badImplementation.calledWithMatch('Insert Rule Error'));
+  t.true(response.boom.badImplementation.calledWithMatch('something bad'));
   t.false(await ruleModel.exists(newRule.name));
   t.is(dbRecords.length, 0);
+  t.false(await t.context.esRulesClient.exists(
+    newRule.name
+  ));
 });
 
-test.serial('POST does not write to DynamoDB or RDS if writing to DynamoDB fails', async (t) => {
+test.serial('post() does not write to Elasticsearch/PostgreSQL if writing to DynamoDB fails', async (t) => {
   const { newRule, testKnex } = t.context;
 
   const failingRulesModel = {
-    exists: () => false,
+    exists: () => Promise.resolve(false),
     create: () => {
       throw new Error('Rule error');
     },
+    delete: () => Promise.resolve(true),
   };
 
   const expressRequest = {
     body: newRule,
     testContext: {
-      dbClient: testKnex,
-      model: failingRulesModel,
+      knex: testKnex,
+      ruleModel: failingRulesModel,
     },
   };
 
@@ -610,23 +619,58 @@ test.serial('POST does not write to DynamoDB or RDS if writing to DynamoDB fails
 
   t.is(dbRecords.length, 0);
   t.false(await ruleModel.exists(newRule.name));
+  t.false(await t.context.esRulesClient.exists(
+    newRule.name
+  ));
+});
+
+test.serial('post() does not write to DynamoDB/PostgreSQL if writing to Elasticsearch fails', async (t) => {
+  const { newRule, testKnex } = t.context;
+
+  const fakeEsClient = {
+    index: () => Promise.reject(new Error('something bad')),
+  };
+
+  const expressRequest = {
+    body: newRule,
+    testContext: {
+      knex: testKnex,
+      esClient: fakeEsClient,
+    },
+  };
+
+  const response = buildFakeExpressResponse();
+
+  await post(expressRequest, response);
+
+  t.true(response.boom.badImplementation.calledWithMatch('something bad'));
+
+  const dbRecords = await t.context.rulePgModel
+    .search(t.context.testKnex, { name: newRule.name });
+
+  t.is(dbRecords.length, 0);
+  t.false(await ruleModel.exists(newRule.name));
+  t.false(await t.context.esRulesClient.exists(
+    newRule.name
+  ));
 });
 
 test('PUT replaces a rule', async (t) => {
-  const putTestRule = {
-    ...t.context.newRule,
-    queueUrl: 'fake-queue-url',
-  };
-  t.truthy(putTestRule.queueUrl);
-  const postgresRule = await translateApiRuleToPostgresRule(putTestRule, t.context.testKnex);
-
-  await t.context.testKnex.transaction(async (trx) => {
-    await t.context.rulePgModel.create(trx, postgresRule);
-    await ruleModel.create(putTestRule, putTestRule.createdAt);
-  });
+  const {
+    originalDynamoRule,
+    originalPgRecord,
+    originalEsRecord,
+  } = await createRuleTestRecords(
+    t.context,
+    {
+      queueUrl: 'fake-queue-url',
+      collection: undefined,
+      provider: undefined,
+    }
+  );
 
   const updateRule = {
-    ...omit(putTestRule, ['queueUrl', 'provider', 'collection']),
+    ...omit(originalDynamoRule, ['queueUrl', 'provider', 'collection']),
     state: 'ENABLED',
     // these timestamps should not get used
     createdAt: Date.now(),
@@ -641,38 +685,41 @@ test('PUT replaces a rule', async (t) => {
     .expect(200);
 
   const actualRule = await ruleModel.get({ name: updateRule.name });
-
   const actualPostgresRule = await t.context.rulePgModel
     .get(t.context.testKnex, { name: updateRule.name });
-  const postgresExpectedRule = await translateApiRuleToPostgresRule(
-    {
-      ...updateRule,
-      createdAt: actualRule.createdAt,
-    },
-    t.context.testKnex
+  const updatedEsRecord = await t.context.esRulesClient.get(
+    originalDynamoRule.name
   );
-  Object.keys(postgresExpectedRule).forEach((key) => {
-    if (postgresExpectedRule[key] === undefined) {
-      postgresExpectedRule[key] = null;
-    }
-  });
 
   t.true(actualRule.updatedAt > updateRule.updatedAt);
   // PG and Dynamo records have the same timestamps
   t.is(actualPostgresRule.created_at.getTime(), actualRule.createdAt);
   t.is(actualPostgresRule.updated_at.getTime(), actualRule.updatedAt);
+  t.is(actualPostgresRule.created_at.getTime(), updatedEsRecord.createdAt);
+  t.is(actualPostgresRule.updated_at.getTime(), updatedEsRecord.updatedAt);
 
   t.like(actualPostgresRule, {
-    ...postgresExpectedRule,
-    created_at: postgresRule.created_at,
+    ...omit(originalPgRecord, ['queue_url']),
+    enabled: true,
+    created_at: new Date(originalDynamoRule.createdAt),
     updated_at: actualPostgresRule.updated_at,
   });
   t.deepEqual(actualRule, {
     // should not contain a queueUrl property
     ...updateRule,
-    createdAt: putTestRule.createdAt,
+    createdAt: originalDynamoRule.createdAt,
     updatedAt: actualRule.updatedAt,
   });
+  t.deepEqual(
+    updatedEsRecord,
+    {
+      ...omit(originalEsRecord, ['queueUrl']),
+      state: 'ENABLED',
+      createdAt: originalDynamoRule.createdAt,
+      updatedAt: actualRule.updatedAt,
+      timestamp: updatedEsRecord.timestamp,
+    }
+  );
 });
 
 test('PUT returns 404 for non-existent rule', async (t) => {
@@ -703,6 +750,196 @@ test('PUT returns 400 for name mismatch between params and payload',
     t.falsy(record);
   });
 
+test('put() does not write to PostgreSQL/Elasticsearch if writing to Dynamo fails', async (t) => {
+  const { testKnex } = t.context;
+  const {
+    originalDynamoRule,
+    originalPgRecord,
+    originalEsRecord,
+  } = await createRuleTestRecords(
+    t.context,
+    {
+      collection: undefined,
+      provider: undefined,
+      queueUrl: 'queue-1',
+    }
+  );
+
+  const fakeRulesModel = {
+    get: () => Promise.resolve(originalDynamoRule),
+    update: () => {
+      throw new Error('something bad');
+    },
+    create: () => Promise.resolve(originalDynamoRule),
+  };
+
+  const updatedRule = {
+    ...omit(originalDynamoRule, ['collection', 'provider']),
+    queueUrl: 'queue-2',
+  };
+
+  const expressRequest = {
+    params: {
+      name: originalDynamoRule.name,
+    },
+    body: updatedRule,
+    testContext: {
+      knex: testKnex,
+      ruleModel: fakeRulesModel,
+    },
+  };
+
+  const response = buildFakeExpressResponse();
+
+  await t.throwsAsync(
+    put(expressRequest, response),
+    { message: 'something bad' }
+  );
+
+  t.deepEqual(
+    await t.context.ruleModel.get({
+      name: originalDynamoRule.name,
+    }),
+    omit(originalDynamoRule, ['provider', 'collection'])
+  );
+  t.deepEqual(
+    await t.context.rulePgModel.get(t.context.testKnex, {
+      name: originalDynamoRule.name,
+    }),
+    originalPgRecord
+  );
+  t.deepEqual(
+    await t.context.esRulesClient.get(
+      originalDynamoRule.name
+    ),
+    originalEsRecord
+  );
+});
+
+test('put() does not write to Dynamo/Elasticsearch if writing to PostgreSQL fails', async (t) => {
+  const { testKnex } = t.context;
+  const {
+    originalDynamoRule,
+    originalPgRecord,
+    originalEsRecord,
+  } = await createRuleTestRecords(
+    t.context,
+    {
+      collection: undefined,
+      provider: undefined,
+      queueUrl: 'queue-1',
+    }
+  );
+
+  const fakerulePgModel = {
+    upsert: () => Promise.reject(new Error('something bad')),
+  };
+
+  const updatedRule = {
+    ...omit(originalDynamoRule, ['collection', 'provider']),
+    queueUrl: 'queue-2',
+  };
+
+  const expressRequest = {
+    params: {
+      name: originalDynamoRule.name,
+    },
+    body: updatedRule,
+    testContext: {
+      knex: testKnex,
+      rulePgModel: fakerulePgModel,
+    },
+  };
+
+  const response = buildFakeExpressResponse();
+
+  await t.throwsAsync(
+    put(expressRequest, response),
+    { message: 'something bad' }
+  );
+
+  t.deepEqual(
+    await t.context.ruleModel.get({
+      name: originalDynamoRule.name,
+    }),
+    omit(originalDynamoRule, ['provider', 'collection'])
+  );
+  t.deepEqual(
+    await t.context.rulePgModel.get(t.context.testKnex, {
+      name: originalDynamoRule.name,
+    }),
+    originalPgRecord
+  );
+  t.deepEqual(
+    await t.context.esRulesClient.get(
+      originalDynamoRule.name
+    ),
+    originalEsRecord
+  );
+});
+
+test('put() does not write to Dynamo/PostgreSQL if writing to Elasticsearch fails', async (t) => {
+  const { testKnex } = t.context;
+  const {
+    originalDynamoRule,
+    originalPgRecord,
+    originalEsRecord,
+  } = await createRuleTestRecords(
+    t.context,
+    {
+      collection: undefined,
+      provider: undefined,
+      queueUrl: 'queue-1',
+    }
+  );
+
+  const fakeEsClient = {
+    index: () => Promise.reject(new Error('something bad')),
+  };
+
+  const updatedRule = {
+    ...omit(originalDynamoRule, ['collection', 'provider']),
+    queueUrl: 'queue-2',
+  };
+
+  const expressRequest = {
+    params: {
+      name: originalDynamoRule.name,
+    },
+    body: updatedRule,
+    testContext: {
+      knex: testKnex,
+      esClient: fakeEsClient,
+    },
+  };
+
+  const response = buildFakeExpressResponse();
+
+  await t.throwsAsync(
+    put(expressRequest, response),
+    { message: 'something bad' }
+  );
+
+  t.deepEqual(
+    await t.context.ruleModel.get({
+      name: originalDynamoRule.name,
+    }),
+    omit(originalDynamoRule, ['provider', 'collection'])
+  );
+  t.deepEqual(
+    await t.context.rulePgModel.get(t.context.testKnex, {
+      name: originalDynamoRule.name,
+    }),
+    originalPgRecord
+  );
+  t.deepEqual(
+    await t.context.esRulesClient.get(
+      originalDynamoRule.name
+    ),
+    originalEsRecord
+  );
+});
+
 test('DELETE deletes a rule', async (t) => {
   const { newRule } = t.context;
 
@@ -725,4 +962,167 @@ test('DELETE deletes a rule', async (t) => {
 
   t.is(dbRecords.length, 0);
   t.is(message, 'Record deleted');
+  t.false(
+    await t.context.esRulesClient.exists(
+      newRule.name
+    )
+  );
+});
+
+test('del() does not remove from PostgreSQL/Elasticsearch if removing from Dynamo fails', async (t) => {
+  const {
+    originalDynamoRule,
+  } = await createRuleTestRecords(
+    t.context,
+    {
+      collection: undefined,
+      provider: undefined,
+    }
+  );
+
+  const fakeRulesModel = {
+    get: () => Promise.resolve(originalDynamoRule),
+    delete: () => {
+      throw new Error('something bad');
+    },
+    create: () => Promise.resolve(true),
+  };
+
+  const expressRequest = {
+    params: {
+      name: originalDynamoRule.name,
+    },
+    testContext: {
+      knex: t.context.testKnex,
+      ruleModel: fakeRulesModel,
+    },
+  };
+
+  const response = buildFakeExpressResponse();
+
+  await t.throwsAsync(
+    del(expressRequest, response),
+    { message: 'something bad' }
+  );
+
+  t.deepEqual(
+    await t.context.ruleModel.get({
+      name: originalDynamoRule.name,
+    }),
+    omit(originalDynamoRule, ['collection', 'provider'])
+  );
+  t.true(
+    await t.context.rulePgModel.exists(t.context.testKnex, {
+      name: originalDynamoRule.name,
+    })
+  );
+  t.true(
+    await t.context.esRulesClient.exists(
+      originalDynamoRule.name
+    )
+  );
+});
+
+test('del() does not remove from Dynamo/Elasticsearch if removing from PostgreSQL fails', async (t) => {
+  const {
+    originalDynamoRule,
+  } = await createRuleTestRecords(
+    t.context,
+    {
+      collection: undefined,
+      provider: undefined,
+    }
+  );
+
+  const fakeRulesPgModel = {
+    delete: () => {
+      throw new Error('something bad');
+    },
+  };
+
+  const expressRequest = {
+    params: {
+      name: originalDynamoRule.name,
+    },
+    testContext: {
+      knex: t.context.testKnex,
+      rulePgModel: fakeRulesPgModel,
+    },
+  };
+
+  const response = buildFakeExpressResponse();
+
+  await t.throwsAsync(
+    del(expressRequest, response),
+    { message: 'something bad' }
+  );
+
+  t.deepEqual(
+    await t.context.ruleModel.get({
+      name: originalDynamoRule.name,
+    }),
+    omit(originalDynamoRule, ['collection', 'provider'])
+  );
+  t.true(
+    await t.context.rulePgModel.exists(t.context.testKnex, {
+      name: originalDynamoRule.name,
+    })
+  );
+  t.true(
+    await t.context.esRulesClient.exists(
+      originalDynamoRule.name
+    )
+  );
+});
+
+test('del() does not remove from Dynamo/PostgreSQL if removing from Elasticsearch fails', async (t) => {
+  const {
+    originalDynamoRule,
+  } = await createRuleTestRecords(
+    t.context,
+    {
+      collection: undefined,
+      provider: undefined,
+    }
+  );
+
+  const fakeEsClient = {
+    delete: () => {
+      throw new Error('something bad');
+    },
+  };
+
+  const expressRequest = {
+    params: {
+      name: originalDynamoRule.name,
+    },
+    testContext: {
+      knex: t.context.testKnex,
+      esClient: fakeEsClient,
+    },
+  };
+
+  const response = buildFakeExpressResponse();
+
+  await t.throwsAsync(
+    del(expressRequest, response),
+    { message: 'something bad' }
+  );
+
+  t.deepEqual(
+    await t.context.ruleModel.get({
+      name: originalDynamoRule.name,
+    }),
+    omit(originalDynamoRule, ['collection', 'provider'])
+  );
+  t.true(
+    await t.context.rulePgModel.exists(t.context.testKnex, {
+      name: originalDynamoRule.name,
+    })
+  );
+  t.true(
+    await t.context.esRulesClient.exists(
+      originalDynamoRule.name
+    )
+  );
 });